requirements.txt
blinker
cachetools
certifi
charset-normalizer
click
colorama
deprecation
dnspython
flask
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
importlib-metadata
itsdangerous
Jinja2
lxml
MarkupSafe
mindee
oauthlib
packaging
pikepdf
Pillow
protobuf
pyasn1
pyasn1-modules
pymongo
pyparsing
pytz
requests
requests-oauthlib
rsa
uritemplate
urllib3
werkzeug
zipp
Gunicorn
